requirements.txt
requests~=2.31.0
beautifulsoup4~=4.12.3
pillow~=10.2.0